requirements.txt
langchain
langchain-community
chromadb
pandas
scikit-learn
fastapi
uvicorn
python-dotenv
openai
sentence-transformers
